Module DkmlDuneDslShow

module I : DkmlDuneDsl.Dune.SYM with type 'a repr = Mustache.Json.t -> Base.Sexp.t

An interpreter of the Dune DSL whose interpreted result is a s-exp of a valid Dune file parameterizable by Mustache JSON configuration.

val plain_hum : ( Mustache.Json.t -> Base.Sexp.t ) list -> string

Show the DSL expression as a plain, human-readable, valid Dune file

val pretty : ( Mustache.Json.t -> Base.Sexp.t ) list -> string

Show the DSL expression as a Dune file with colors if your terminal supports it; do not use for writing Dune files!