Module DkmlDuneDsl.Dune

dune files are the main part of Dune. They are used to describe libraries, executables, tests, and everything Dune needs to know about.

Dune files

dune files are the main part of Dune. They are used to describe libraries, executables, tests, and everything Dune needs to know about.

The official documentation is on the Dune website.

module type SYM = sig ... end

The module type for an embedded domain specific language (eDSL) that describes a Dune file.